Fixed discounts.

+++ final/src/main/java/ua/epam/spring/hometask/service/DiscountService.java
@@ -23,5 +23,5 @@ public interface DiscountService {
      * @param numberOfTickets Number of tickets that user buys
      * @return discount value from 0 to 100
      */
-    byte getDiscount(@Nullable User user, @Nonnull Event event, @Nonnull LocalDateTime airDateTime, long numberOfTickets);
+    double getDiscount(@Nullable User user, @Nonnull Event event, @Nonnull LocalDateTime airDateTime, long numberOfTickets);
 }
